Moved Mythic Plugin implementation out of the separate module.
