+++ plugin-compatibility/plugin-compatibility-artifact/src/main/java/me/wolfyscript/utilities/compatibility/plugins/mmoitems/MMOItemsRefImpl.java
@@ -22,15 +22,14 @@ import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.SerializerProvider;
 import io.lumine.mythic.lib.api.item.NBTItem;
+import java.io.IOException;
+import java.util.Objects;
 import me.wolfyscript.utilities.api.inventory.custom_items.references.APIReference;
 import net.Indyuce.mmoitems.MMOItems;
 import net.Indyuce.mmoitems.api.Type;
 import net.Indyuce.mmoitems.api.item.mmoitem.MMOItem;
 import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.Nullable;
-
-import java.io.IOException;
-import java.util.Objects;
 
 /**
  * Links to MMOItems and saves the specified {@link Type} and Name of the item.
