+++ plugin-compatibility/plugin-compatibility-artifact/src/main/java/me/wolfyscript/utilities/compatibility/plugins/mythicmobs/MythicMobs5RefImpl.java
