+++ plugin-compatibility/plugin-compatibility-artifact/src/main/java/me/wolfyscript/utilities/compatibility/plugins/mythicmobs/MythicMobsRefImpl.java
