+++ plugin-compatibility/pom.xml
@@ -14,7 +14,6 @@
 
     <modules>
         <module>mmoitems</module>
-        <module>oraxen</module>
         <module>mythicmobs</module>
         <module>plugin-compatibility-artifact</module>
     </modules>
